+++ tests/failures.rs
@@ -1,5 +1,8 @@
 use crate::utils::get_test_path;
-use cargo_tarpaulin::{config::Config, errors::RunError};
+use cargo_tarpaulin::{
+    config::{Config, Mode},
+    errors::RunError,
+};
 use cargo_tarpaulin::{launch_tarpaulin, run};
 use std::env;
 
@@ -37,4 +40,22 @@ fn error_if_test_fails() {
     } else {
         panic!("Expected a TestCompile error");
     }
+}
+
+#[test]
+fn issue_610() {
+    let mut config = Config::default();
+    let test_dir = get_test_path("issue_610");
+    config.test_names.insert("foo".to_string());
+    env::set_current_dir(&test_dir).unwrap();
+    config.manifest = test_dir;
+    config.manifest.push("Cargo.toml");
+
+    let result = run(&[config.clone()]);
+    assert!(result.is_ok());
+
+    config.test_names.clear();
+    config.command = Mode::Build;
+    let result = run(&[config]);
+    assert!(result.is_err());
 }