+++ CHANGELOG.md
@@ -3,10 +3,13 @@
 From 2019 onwards, all notable changes to tarpaulin will be documented in this
 file.
 
-## [Unreleased]
+## [0.16.0] - 2020-11-02
 ### Added
+- `--command` option to build and run a binary for testing CLI apps
 
 ### Changed
+- Make `--run-types` and `--out` case insensitive
+- Filter executables on command not run type to fix #610
 
 ### Removed